requirements.txt
cloudinary>=1.44.1
fastapi>=0.135.1
google-generativeai>=0.8.6
httpx>=0.28.1
langchain>=1.2.12
langchain-community>=0.4.1
langchain-core>=1.2.19
langchain-google-genai>=4.2.1
langchain-groq>=1.1.2
langchain-openai>=1.1.11
langgraph>=1.1.2
mixedbread>=0.49.0
numpy>=2.4.3
pandas>=3.0.1
pillow>=12.1.1
psycopg2-binary>=2.9.11
pydantic>=2.12.5
pydantic-settings>=2.13.1
pypdf2>=3.0.1
python-docx>=1.2.0
python-dotenv>=1.2.2
python-multipart>=0.0.22
python-pptx>=1.0.2
sqlalchemy>=2.0.48
supabase>=2.28.1
tavily-python>=0.7.23
torch>=2.11.0
uvicorn>=0.42.0
vecs>=0.4.5
watchfiles>=1.1.1
